+++ src/sync/events/data/index.ts
@@ -7,6 +7,7 @@ import * as weth from "@/events-sync/data/weth";
 import * as element from "@/events-sync/data/element";
 import * as foundation from "@/events-sync/data/foundation";
 import * as looksRare from "@/events-sync/data/looks-rare";
+import * as nouns from "@/events-sync/data/nouns";
 import * as openDao from "@/events-sync/data/opendao";
 import * as quixotic from "@/events-sync/data/quixotic";
 import * as rarible from "@/events-sync/data/rarible";
@@ -58,7 +59,8 @@ export type EventDataKind =
   | "element-erc721-buy-order-filled"
   | "element-erc1155-sell-order-filled"
   | "element-erc1155-buy-order-filled"
-  | "quixotic-order-filled";
+  | "quixotic-order-filled"
+  | "nouns-auction-settled";
 
 export type EventData = {
   kind: EventDataKind;
@@ -108,6 +110,7 @@ export const getEventData = (eventDataKinds: EventDataKind[] | undefined) => {
       element.erc1155BuyOrderFilled,
       element.erc1155SellOrderFilled,
       quixotic.orderFulfilled,
+      nouns.auctionSettled,
     ];
   } else {
     return (
@@ -196,6 +199,8 @@ const internalGetEventData = (kind: EventDataKind): EventData | undefined => {
       return element.erc1155BuyOrderFilled;
     case "quixotic-order-filled":
       return quixotic.orderFulfilled;
+    case "nouns-auction-settled":
+      return nouns.auctionSettled;
     default:
       return undefined;
   }